autogole-api/requirements.txt
psutil
requests
pyyaml
grafana-client
